requirements.txt
Pillow==5.1.0
pyftpdlib==1.5.3
PyQt5==5.10.1
pytz==2018.3
sip==4.19.8
